validacao casso com zod

+++ controllers/casosController.js
@@ -1,84 +1,105 @@
 const casosRepository = require('../repositories/casosRepository');
 const agentesRepository = require('../repositories/agentesRepository');
-const ApiError = require('../utils/ApiError');
-const { getAgenteOrThrowApiError } = require('./agentesController');
-
-function getCasoOrThrowApiError(id) {
-    const caso = casosRepository.findById(id);
-    if (!caso) {
-        throw new ApiError(404, `Não foi possível encontrar o caso de Id: ${id}`);
-    }
-    return caso;
-}
 
 function getAllCasos(req, res) {
     const agente_id = req.query.agente_id;
     const status = req.query.status;
 
     let casos = casosRepository.findAll();
-    if (agente_id) {
-        casos = casos.filter((caso) => caso.agente_id === agente_id);
-        if (casos.length === 0) {
-            throw new ApiError(404, `Nenhum caso foi encontrado para o agente de Id: ${agente_id}`);
-        }
-    }
     if (status) {
         casos = casos.filter((caso) => caso.status === status);
         if (casos.length === 0) {
-            throw new ApiError(404, `Nenhum caso foi encontrado com o status: ${status}`);
+            return res.status(404).json({
+                message: `Não foi possível encontrar casos com o status: ${status}.`,
+            });
         }
     }
-
+    if (agente_id) {
+        casos = casos.filter((caso) => caso.agente_id === agente_id);
+        if (casos.length === 0) {
+            return res.status(404).json({
+                message: `Nenhum caso foi encontrado para o agente de Id: ${agente_id}`,
+            });
+        }
+    }
     res.status(200).json(casos);
 }
 
 function getCasoById(req, res) {
     const id = req.params.id;
-    const caso = getCasoOrThrowApiError(id);
+    const caso = casosRepository.findById(id);
+    if (!caso) {
+        return res.status(404).json({
+            message: `Não foi possível encontrar o caso de Id: ${id}.`,
+        });
+    }
     res.status(200).json(caso);
 }
 
 function getAgenteByCaso(req, res) {
-    const caso_id = req.params.caso_id;
-    const caso = getCasoOrThrowApiError(caso_id);
-
+    const caso_id = req.params.id;
+    const caso = casosRepository.findById(caso_id);
+    if (!caso) {
+        return res.status(404).json({
+            message: `Não foi possível encontrar o caso de Id: ${caso_id}.`,
+        });
+    }
     const agente = agentesRepository.findById(caso.agente_id);
+    if (!agente) {
+        return res.status(404).json({
+            message: `O caso de Id: ${caso_id} não possui um agente associado a ele.`,
+        });
+    }
     res.status(200).json(agente);
 }
 
 function searchCasos(req, res) {
-    const search = req.query.q?.trim();
+    const search = req.query.q?.trim().toLowerCase();
     if (!search) {
-        throw new ApiError(404, "Parâmetro de pesquisa 'q' não encontrado");
+        return res.status(404).json({ message: "Parâmetro de pesquisa 'q' não encontrado" });
     }
 
     const searchedCasos = casosRepository.search(search);
     if (searchedCasos.length === 0) {
-        throw new ApiError(404, 'Não foi possível encontrar casos que correspondam à pesquisa');
+        return res.status(404).json({
+            message: `Não foi possível encontrar casos que correspondam à pesquisa: ${search}.`,
+        });
     }
     res.status(200).send(searchedCasos);
 }
 
 function postCaso(req, res) {
     const { titulo, descricao, status, agente_id } = req.body;
-    getAgenteOrThrowApiError(agente_id);
+    if (!agentesRepository.findById(agente_id)) {
+        return res.status(404).json({
+            message: `Não foi possível encontrar o agente de Id: ${agente_id}.`,
+        });
+    }
 
-    const createdCasoData = {
+    const newCasoData = {
         titulo,
         descricao,
         status,
         agente_id,
     };
-    const createdCaso = casosRepository.create(createdCasoData);
+    const createdCaso = casosRepository.create(newCasoData);
     res.status(201).json(createdCaso);
 }
 
 function updateCaso(req, res) {
     const id = req.params.id;
-    getCasoOrThrowApiError(id);
+    if (!casosRepository.findBy(id)) {
+        return res.status(404).json({
+            message: `Não foi possível encontrar o caso de Id: ${id}.`,
+        });
+    }
 
     const { titulo, descricao, status, agente_id } = req.body;
-    getAgenteOrThrowApiError(agente_id);
+    if (!agentesRepository.findById(agente_id)) {
+        return res.status(404).json({
+            message: `Não foi possível encontrar o agente de Id: ${agente_id}.`,
+        });
+    }
 
     const updatedCasoData = {
         titulo,
@@ -92,10 +113,19 @@ function updateCaso(req, res) {
 
 function patchCaso(req, res) {
     const id = req.params.id;
-    const caso = getCasoOrThrowApiError(id);
+    const caso = casosRepository.findById(id);
+    if (!caso) {
+        return res.status(404).json({
+            message: `Não foi possível encontrar o caso de Id: ${id}.`,
+        });
+    }
 
     const { titulo, descricao, status, agente_id } = req.body;
-    if (agente_id !== undefined) getAgenteOrThrowApiError(agente_id);
+    if (!agentesRepository.findById(agente_id)) {
+        return res.status(404).json({
+            message: `Não foi possível encontrar o agente de Id: ${agente_id}.`,
+        });
+    }
 
     const patchedCasoData = {
         titulo: titulo ?? caso.titulo,
@@ -109,7 +139,12 @@ function patchCaso(req, res) {
 
 function deleteCaso(req, res) {
     const id = req.params.id;
-    getCasoOrThrowApiError(id);
+    const caso = casosRepository.findById(id);
+    if (!caso) {
+        return res.status(404).json({
+            message: `Não foi possível encontrar o caso de Id: ${id}.`,
+        });
+    }
 
     casosRepository.remove(id);
     res.status(204).send();
